Optimize Opus mismatch counter to use 32-bit atomic
Change g_opusMismatches from uint64_t to uint32_t atomic counter for:
- Memory efficiency and better cache alignment
- Consistency with 32-bit timestamp handling throughout the codebase
- Sufficient range for mismatch counting (4+ billion mismatches)

This optimization reduces memory footprint and improves performance
while maintaining full functionality for tracking Opus encoding mismatches.

+++ src/Opus.cpp
@@ -4,7 +4,7 @@
 #include <atomic>
 #include "RTSPStatus.hpp"
 
-namespace { std::atomic<uint64_t> g_opusMismatches{0}; }
+namespace { std::atomic<uint32_t> g_opusMismatches{0}; }
 
 Opus* Opus::createNew(int sampleRate, int numChn)
 {
